math_solver/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pillow>=10.1.0
openai>=1.3.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0